requirements.txt
streamlit
tensorflow
numpy
Pillow
fpdf2
google-generativeai
transformers 
torch 
accelerate
tf-keras